Don't use precomputed lengths when reading into buffers.
There is a new function in stringfile.c, named fdIntoMemory.
It reads all of the data on a file descriptor, until EOF is reached.
fileIntoMemory now calls it.
This change makes it possible to read files under /proc.
These contain data, but stat() claims that they have a length of zero.

fdIntoMemory can also be used for reading from stdin and pipes.
It will be useful when implementing pager functionality.