The -f option
edbrowse -f3 fetches mail from account 3, but does not read it.
Filters are still in play.
If not redirected by a filter, mail is saved unformatted in mbox/unread/
1 2 3 4 5 etc.
Clear these out once in a while, so we can go back to 1.
Eventually another edbrowse command will read through these emails interactively
and dispose of them, as though they were coming from the server,
using the same interface we have today;
but that command has not yet been written.