Restructure the email interface to separate -f and -m
-f is now fetchmail, -m reads the mail in mbox/unread.
-fm is shorthand to do both.
Fetch is done first, then the scan through the emails;
so you're not under the gun on what to do with each email,
before the pop3 server times out.
-fm# specify an account number to fetch your mail from.
-fm fetch mail from all accounts,
excluding duplicates, and excluding accounts with "nofetch" in the descriptor.
Filtering takes place in the -m side, so any fetchmail client could be used,
as long as some follow-on shell script places
the mail files in the unread directory,
and follows our numbering convention  1 2 3 4 etc.