readline: don't append trailing space when completing.
edbrowse commands such as r and w interpret the space character
literally.  Let's say that we want to read a file named foo,
and we're using tab completion to complete the partial filename.
With readline's default behavior,
r fo<tab>
becomes
r foo
(note the trailing space).
We try to read a file named "foo ", which is not what the user wanted.
The fix is trivial.