Fix id=submit bug, as illustrated by stargpage.com.
The id=submit tag overwrites the form.submit() function.
It's not clear that the id=foo should create form.foo at all,
so for now I comment it out,
with an explanation of why.
We'll see if some other website needs me to create form.foo from id=foo.
I hope it is not some wacky logic like create the link
iff the moon is full.