Remove AutoCompartment from a few routines.
This patch and the next couple shouldn't change the behavior of edbrowse.
In some functions I can remove the rooted objects altogether,
so that they run fine whether js is active or not.
I use to have shorthand variables like e = topTag->jv.
That is no longer shorthand, because I have to root the variable,
and I also have to set the compartment.
So it is often easier just to refer to topTag->jv.
Then the additional advantage:
the function doesn't blow up if cw->jss is null.
As of this push all functions in html.cpp up to htmlImage
are safe to call whether jss is null or not.