Keep the tags in one vector, not two separate structures in parallel.
Recall that tags were appended to a linked list as the page was parsed,
then put in a fixed array under the current window after parse.
This was confusing; it was hard to know
which structure to look in for the tags.
if parsePage { look here } else { look there }.   Yuck!
Beyond this, it assumes new tags will never be created once
the page is parsed; we know that is not true.
Thus all tags are in a growing vector from the get-go,
which still allows for easy random access when you push a button
on the page.
Since vector is a c++ construct, eb.h simply has to call it void *.
That's ok though; nobody uses it but html.cpp, so one cast and we're there.