Add function get_property_int, like the other get_property functions.
This is used to get array.length, or other things you know are ints.
I also took out some of the failure legs in these functions.
Javascript should not shut down entirely
just because an attribute o.member is not there.
That does not indicate a js failure, just a missing member.

Also the outline of the rebuildSelectors function,
which is not called yet.