Several small changes to make the new fetchmail layer work through curl.
These are just little fixes, the most important
is to count messages based on the response from list,
which is a little more than just counting lines.

This layer doesn't work with curl 7.21, but does with curl 7.36.
I don't know where in between is the cutoff.
I called it 7.29, just to make a guess.