The first implementation of object.parentNode.
This is probably not right but it's a start.
I can't find good docs on this.
Basically it steps up the tree of your web page,
from option to selector to form to paragraph to body to document, for example.
Set debug 5 to see the parent links built.
This fixes the error in amazon.com, though another error crops up
on the very next line, so on we go.