Delete tcp.c tcp.h, we don't need them any more.
Allow for an empty subject, becoming no subject in the generated email.
For a proper subject,
pass the subject directly to curl, hoping it will
encode as ascii, or whatever it must do to transmit.
I tested this with a nonascii char in subject and it came across ok.

In messages.h, mark as unused the messages we don't use any more.
There are 52 of them.
Most of these were my own as I implemented pop3 and smtp from the ground up.
Now using curl error messages, which are much fewer.
Other messages aren't needed because we can now do
the things they said we couldn't do, like having no subject.

Comment on curl pop3 smtp in CHANGES.

Adjust COPYING and the users guide, don't need the openssl special exception
in the description of copyright and gpl.
Curl brings in ssl, and it's now curl's problem
to present the special openssl legal copyright crap.