nodeName attribute of most of the dom objects indicates the html tag that created it.
I put this in the domLink() function, that does most of the work
of creating a js object for the html tag.
That function is getting hairy.

,.
The option constructor sets nodeName = "OPTION"
when a new option is made out of thin air,
probably I should be doing the same for other dom options that are created
dynamically and not from html.