Put null on the end of the blacklist file.
I read the blacklist file into memory, then parsed it line by line,
stopping at null, except I didn't null terminate the string
after I read the file into memory.
So the parsing of the file just marched on and on, until it hit a 0.
Very rarely that would corrupt the heap and lead to a malloc free
error farther down the line.
This is a one line fix, putting '\0' on the end.