Remove reference to libcrypto, it is obsolete.
This was here when we called the openssl routines directly, for https,
and then we had to link with crypto, but now,
we don't call openssl at all, curl does it for us, and,
libcrypto isn't there any more, gone in the newer systems.
So crypto needs to be deleted for edbrowse to build on newer systems,
and no need to reference it even on the old systems,
cause curl brings it in for us.