Use the title as the description of the hyperlink created by the A command.
This change suggested by Tyler Spivey.
A bit more complicated than first thought because:
A) title does not need to be stripped and chopped
as is done with the url.
B) We need the original form of the title for the link,
before andTranslate was run. Example www.horror.com
C) I also wanted to update the documentation.