Remove messages from messages.c that are not being used.
These have been removed or replaced in English, but not in other languages.
So run edbrowse in French, and you still get the old, wrong message.
Put in 0, because I don't know the french equivalent of the new message.
It has not yet been translated.
If edbrowse finds a 0 there it reverts back to English.
That's better than nothing.
At least you get the right message, albeit in the wrong language.
I'll ask my friends to translate more messages after the dust settles.