Use wordexp(3) to expand variables and stars on filenames.
Also use wordexp to gather all the files under a directory
in a directory listing such as edbrowse /
This was a bit tricky, since I had to run it twice, for * and .*
The latter to see the hidden files.
This is all a bit experimental.
I left the old code in, for now, under #if 0,
we can chop it out when we have a bit more confidence in the new system.