Push js foo.value = string changes onto a queue, then apply them.
This mechanism was already there, but sometimes I would use it,
and sometimes not - instead updating the edbrowse text buffer immediately.
It is more consistent to always do it through the queue.
It is just another side effect of running js.
This should not change at all the way edbrowse runs today.