Queue up the innerHTML and innerText side effects from js.
These are not applied yet, just put into the queue.
For simplicity I'm using the same queue as the input.value changes.
Perhaps other js side effects can be put into this queue as well.
I'm using major to indicate the item,
for now the same letter as the switch statement in side effects in ebjs.c.
v is an assignment to input.value
i is innerHTML or innerText, with minor set to h or t