Set the base url for document.write() after the page is parsed.
I say that, but don't really know what the base should be.
The original url, or the last <base> tag?
Also add some jsrt code to run a dynamic script
out of the onload function, which does document.write, which creates a hyperlink,
which tests the base url.
Should resolve to
http://yyy.zzz.com/custer