Create an internal invisible tag for almost every html tag.
This will support innerHTML, which can appear in almost any tag,
but usually <div>.
All these new invisible internal tags required considerable changes
in the text formatting software, which needed to be changed anyways.
It is simpler now and more consistent.
The key is not to advance the column number for the invisible tags,
and then the formatting doesn't get screwed up.