Handle percent-encoding for both types of data URIs.
The spec is vague on this, but apparently,
the equals signs in the base64 encoding may be URL-encoded,
so we have to URL decode before passing the string to base64Decode.
Google Chrome accepts unencoded equals signs in data:
URIs, so presumably we should be ready to expect both encoded and raw.