Go back to opendir readdir for the directory scan.
Yeah I know, a bit of thrashing here. This is how it looked 4 versions ago.
But this was profitable thrashing; envFile() really needs to use glob().
My home grown globbing code wasn't very good.
Did not expand stars in the middle of a path etc.
But scanning a directory, gees, that's not hard. I can do that.
And it fixes Tyler's bug, wherein broken symlinks were not shown,
and it's actually less code then setting up for glob(),
and it allows us to sort the files as we wish, which we might want to
do based on locale, rather than simple strcmp.
Not sure about that one yet.
I didn't delete the globProtect routines, in case, God forbid,
we have to go back to them again.
If everyone is happy with this incarnation then I'll delete globProtect() later.