Fix bug in javascript document.cookie
This is a specialized variable with getter and setter.
I tried to do everything with the setter, but the setter cannot change
the value being set, it can only do side effects to setting that variable,
at least in spider monkey js, as far as I can tell.
So I had to put in a getter to pull back the complete cookie string.
Also added tests for this in jsrt.
Look for document.cookie in that file.
I haven't tested document.cookie when persistent cookies
are preloaded based on the url.
This is hard to do with jsrt as that is not a proper url,
and I don't think jsrt persistent cookies would work in your cookie jar.