Make the percentURL() routine a tad closer to RFC3986, but it's still not right.
It assumes some of the reserved chars are performing their reserved functions
and does not escape those,
but honestly it really doesn't know and can't tell for sure.
It's usually right.
This change escapes [] which were not escaped before,
but must be for imap urls that use [Gmail] in the path.