Fix bug resetting a textarea in a form, see buffer 2 in jsrt.
This broke when we said a file with no name can still
deny the quit command if its data has been updated.
That's fine interactive, but in this case we really want to reset the data,
even if the user has changed it.
The same goes for innerText = string, which was also broken and is now fixed.
See the <Rewrite> option in jsrt.