Fix space + bug caused by url post data being encoded twice.
This is for form data being sent to a server.
Turning space into + is legal, but if encoded again,
+ becomes %2b, which is a literal + and no longer a space indicator.
Post data can be encoded multiple times, and I want this to be safe,
so I got rid of the space + translation and go straight to %20,
which is unchanged by further encodings.
This fixes the bug for google search
cat dog becomes cat%20dog and reverts back to cat dog in the search bar.
Not so for the wikipedia search.
Put in cat dog and set db3 and search,
and notice how it starts out as cat%20dog but then becomes cat%2bdog,
as though it went through cat+dog in between.
Does wikipedia have the same bug that I had?
Maybe it's an easy bug to write.