Pass the html node to domLink() as a parameter, rather than a global variable.
Globals are usually a bad idea,
and with this change that global variable goes away.
doesn't need to be global for any other reason.
Some of the calls in html.c get a little more ugly with the extra parameter,
splitting across two lines, but those calls are going away anyways.
It's all going to move to render(), hopefully cleaned up a bit along the way.