Don't allow html or js to set the filename in a form, thus uploading a file of its choosing.
This is a security fix!!
Prevent tags like
<input type=file value=/etc/passwd>
whereupon js could submit the form via script or onload function
and send up your password file.
Even more frightening is .ssh/id_rsa
Also prevent
form.upfile.value = "/etc/passwd";
This can still happen in js, but there are no side effects,
so the change does not make it back to edbrowse
and that file will not be uploaded.
Also don't dip into js for the value, use the tag value only.
The upshot is: edbrowse, or any other browser for that matter,
should only send up a file that you have typed in, exactly as you typed it.