Fix setters, and also fix innerHTML.
I can't use just a setter, can't figure it out.
Have to have both getter and setter.
I have to store the value in another "raw" property, and a getter fetches that value, and a setter sets it and performs the side effects.
It's a little indirect compared to moz js, but it works, and I guess it's intuitive once you get use to it.