main: run arguments as function if first char is '<'
This lets someone run "edbrowse '<mw foo'" or other functions
defined in their ebrc, and have it displayed in a context.

The quotes are necessary because a single word (argv item) is used
for the function, "edbrowse '<mw' foo" would run just mw without
argument and then open foo as second file.