Debug alert watches for ^c interrupt.
If it's an infinite loop, or crazy long, and I can't stop it, I can at least rerun it with uvw trace, and the debug outputs check for ^c, and then I can stop it.
I did this to help diagnose http://songmeanings.com/songs/view/3530822107858535238/
which I thought was infinite but is only god-awful slow.