<base> tag carries the url over to window.location and document.location if the file jslocal is present. This provides a more faithful simulation under the snapshot feature. The whole page looks like it runs from the base.
jquery and ajax assume the page runs from window.location, so the snapshot run doesn't work properly if window.location is some local file.
It has to be the same url as <base>.
The implementation was a bit tricky because I didn't want to trigger the redirection machinery that usually kicks in when you change window.location.