Fix URL: support when transforming gopher menus to HTML
Previously the URL was encoded and then a host etc added because the encoding
of the path was done prior to checking for the string "URL:" in the path.
This commit moves encoding the path after the check and avoids doing so if it
contains a full URL (e.g. to an HTTP resource).