Implement asynchronous xhr.
All these threads and asynchronous actions are fun, but only speed up the browse by a second  or two, and then you're annoyed cause it says so many lines are updated,
and it's distracting, so might not even be worth it.
Better to wait 3 more seconds and have the actual web page in its entirety.
So jsbg is off by default.
I'll have to play with this for a while.