Fix duktape linking order in the GNUmakefile
Apparently it's not guaranteed that the Duktape shared library is linked
against the math functions.  In fact, on Debian, it isn't.  This means that,
if the math functions are in the wrong place in the linking list, the linker
complains that, on linking in Duktape, it can't find them.  Apparently.  the
Correct place for Duktape is prior to -lm, so move it to before LDLIBS in
the GNUmakefile.

Note that I'm not touching the other Makefile since I can't test it
currently.