Get rid of docwrap for now. It didn't fix all the problems with mozzila, it's complicated, it's ugly, and it seems unnecessary in duktape and in quick.
I just commented it out, in case we want to resurrect it or learn from it.
startwindow I actually had to change, but the docwrap lines will always be preserved in github.
This and other changes means edbrowsesm will no longer build, but I'm almost committed to go down a different path now.