Change fetch limit while viewing envelopes.
I didn't see any point to this, as you would set the fetch limit at the top, before you descended into a folder,
however, you might want to just fetch the last 10, out of 10 thousand, then change the limit to 1000, then search for all the mail from somebody, or with a subject, etc.
So there really is a rationale for this.