Patch for edbrowse
In this patch, I moved everything that’s currently done in make-helper.sh into the makefile itself. I also modified the format of make-helper.sh a bit, added subcommands to get pkg-config includes and libs, and cleaned up the makefile a bit. I also modified the install command to work with BSD as well as GNU install, and copy over the users guide (English only).