Ignore trailing / when comparing two urls, / is optional when the resource is a directory.
This allows us to change hash instead of redirect, when one of the urls has trailing / and the other does not.
This is more efficient. It is seen in www.radiocaroline.co.uk
which redirects to www.radiocaroline.co.uk/home.html
Often the server will 301 redirect and put the / on for us but not this time.
We have the page without the slash, and browsing, and then it redirects to set the hash.
Note, this code assumes both urls are valid.
example.com/snork and example.com/snork/ are not the same if snork is a file; the latter is an http error.