Watch for window or document object == null.
This is a corner case, described in the comments, but even that case shouldn't happen.
When checking for css visibility I should be working within the frame being rendered, not the frame at the top of the page,
so we really need to fix that some day before it causes more trouble.