Be really aggressive about setting close-on-exec.
I did this for libcurl sockets a long long time ago.
But there are tons of other places where we aren't, and should
be, using close-on-exec.  For instance, if you shell out to
something, there is no earthly reason why the kiddo and
whatever descendents it may have should have FDs pointing
at edbrowse's cache control files and suchlike.
So let's be really really aggressive.

fopen has an e mode for this; open has O_CLOEXEC.
Note that fopen seems to be really picky about where the e
goes in the modestring.  Apparently it must always come at
the end.

This was a bit of a naive search-and-replace.  Not
particularly well-tested, though things appear fine.